requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug 
PyJWT
pymysql
cryptography